svm: Scale data before splitting into training and test sets
Dataset needs to be scaled as a whole set, before it is divided into
training and test subsets. Otherwise the SVC is tested on a data that is
scaled in a different manner (i.e. different outliers could have
affected it) and model performance suffers significantly.

This improvement alone increases score of the model from 0.7 to 0.9.